backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.3.0
python-dotenv==1.0.1
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
pydantic==2.6.0
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==4.3.1
sentence-transformers>=2.3.1

